long test separation (Fix #156)
