+++ CMakeLists.txt
@@ -67,6 +67,8 @@ IF (EXISTS ${PROJECT_SOURCE_DIR}/tests)
         MESSAGE (STATUS "Nosestests exists")
         ENABLE_TESTING ()
         ADD_SUBDIRECTORY (tests)
+        ADD_CUSTOM_TARGET (test-unit COMMAND ${CMAKE_CTEST_COMMAND} WORKING_DIRECTORY tests/unit)
+        ADD_CUSTOM_TARGET (test-long COMMAND ${CMAKE_CTEST_COMMAND} WORKING_DIRECTORY tests/long)
     ELSE ()
         MESSAGE (WARNING "Nosetests is missing! - no tests available")
     ENDIF ()
