+++ tests/CMakeLists.txt
@@ -1,4 +1,2 @@
-FILE (GLOB files test_*.py)
-FOREACH (file ${files})
-    ADD_TEST(${file} nosetests-3.4 ${file} -w ${PROJECT_SOURCE_DIR})
-ENDFOREACH ()
+ADD_SUBDIRECTORY (unit)
+ADD_SUBDIRECTORY (long)
